chore: rename agentLogos to AgentLogo and co-locate gemini logo
Rename the shared agentLogos module to AgentLogo.tsx and move the gemini SVG
into the AgentPromptButton/logos directory alongside the other brand marks, so
all agent logos live in one place rather than reaching into the onboarding
scene.

Generated-By: PostHog Code
Task-Id: 693b0508-10d0-4fd6-ae97-8d8a2bbb5b05
